src/main/java/com/lj/service/impl/ProductCategoryServiceImpl.java
package com.lj.service.impl;

import com.lj.entity.ProductCategory;
import com.lj.repository.ProductCategoryRepository;
import com.lj.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by lj0782 on 2017/11/1.
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {
    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory findOne(Integer id) {
        return repository.findOne(id);
    }
}
